refactor(commands): replace deep imports with public API imports (#2444)
Replace all deep cross-module imports in src/vibe3/commands/*.py with
top-level public API imports. The change is mechanical: rewrite
`from vibe3.<module>.<submodule> import <symbol>` to
`from vibe3.<module> import <symbol>`.

Also update roles/__init__.py __getattr__ return type from object to Any
for mypy compatibility with lazy imports, and update test patch targets
to match the new import paths.

Affects 21 command files (54 import lines updated) + 6 test files.
