+++ src/vibe3/commands/check_support.py
@@ -104,7 +104,7 @@ def execute_check_mode(
 
     if mode == "clean_branch":
         # Use dedicated cleanup service for --clean-branch
-        from vibe3.services.check_cleanup_service import CheckCleanupService
+        from vibe3.services import CheckCleanupService
 
         cleanup_service = CheckCleanupService(
             store=service.store,
