+++ src/vibe3/commands/inspect_base.py
@@ -56,9 +56,8 @@ def register(app: typer.Typer) -> None:
             vibe inspect base origin/develop # Compare current branch vs origin/develop
             vibe inspect base main           # Compare current branch vs origin/main
         """
-        from vibe3.clients.git_client import GitClient
-        from vibe3.clients.github_client import GitHubClient
-        from vibe3.config.loader import get_config
+        from vibe3.clients import GitClient, GitHubClient
+        from vibe3.config import get_config
         from vibe3.models.change_source import BranchSource
         from vibe3.utils.git_helpers import get_current_branch
