+++ tests/vibe3/commands/test_flow_status_remote_timeline.py
@@ -42,9 +42,7 @@ def test_remote_no_local_db_uses_remote_timeline() -> None:
     )
 
     with patch("vibe3.commands.flow_status.FlowService") as mock_service_class:
-        with patch(
-            "vibe3.services.flow_status_resolver.FlowStatusResolver"
-        ) as mock_resolver_class:
+        with patch("vibe3.services.FlowStatusResolver") as mock_resolver_class:
             with patch(
                 "vibe3.commands.flow_status.FlowProjectionService"
             ) as mock_projection_class:
@@ -108,9 +106,7 @@ def test_remote_ignores_stale_local_events() -> None:
     )
 
     with patch("vibe3.commands.flow_status.FlowService") as mock_service_class:
-        with patch(
-            "vibe3.services.flow_status_resolver.FlowStatusResolver"
-        ) as mock_resolver_class:
+        with patch("vibe3.services.FlowStatusResolver") as mock_resolver_class:
             with patch(
                 "vibe3.commands.flow_status.FlowProjectionService"
             ) as mock_projection_class:
@@ -169,9 +165,7 @@ def test_non_remote_uses_local_events() -> None:
     )
 
     with patch("vibe3.commands.flow_status.FlowService") as mock_service_class:
-        with patch(
-            "vibe3.services.flow_status_resolver.FlowStatusResolver"
-        ) as mock_resolver_class:
+        with patch("vibe3.services.FlowStatusResolver") as mock_resolver_class:
             with patch(
                 "vibe3.commands.flow_status.FlowProjectionService"
             ) as mock_projection_class:
@@ -244,9 +238,7 @@ def test_remote_json_output_includes_timeline() -> None:
     )
 
     with patch("vibe3.commands.flow_status.FlowService") as mock_service_class:
-        with patch(
-            "vibe3.services.flow_status_resolver.FlowStatusResolver"
-        ) as mock_resolver_class:
+        with patch("vibe3.services.FlowStatusResolver") as mock_resolver_class:
             with patch(
                 "vibe3.commands.flow_status.FlowProjectionService"
             ) as mock_projection_class:
@@ -300,9 +292,7 @@ def test_remote_yaml_output_includes_timeline() -> None:
     )
 
     with patch("vibe3.commands.flow_status.FlowService") as mock_service_class:
-        with patch(
-            "vibe3.services.flow_status_resolver.FlowStatusResolver"
-        ) as mock_resolver_class:
+        with patch("vibe3.services.FlowStatusResolver") as mock_resolver_class:
             with patch(
                 "vibe3.commands.flow_status.FlowProjectionService"
             ) as mock_projection_class:
